+++ init.php
@@ -13,6 +13,7 @@ define( 'TRAFFIC_PRODUCT_SHORTNAME', 'Traffic' );
 define( 'TRAFFIC_PRODUCT_ABBREVIATION', 'traffic' );
 define( 'TRAFFIC_SLUG', 'traffic' );
 define( 'TRAFFIC_VERSION', '2.0.0-dev0' );
+define( 'TRAFFIC_API_VERSION', '1' );
 define( 'TRAFFIC_CODENAME', '"-"' );
 
 define( 'TRAFFIC_CDN_AVAILABLE', true );